+++ corekit/src/main/java/com/midtrans/sdk/corekit/core/snap/model/pay/request/CustomerDetailRequest.java
@@ -1,4 +1,4 @@
-package com.midtrans.sdk.corekit.core.snap.model.pay.request.va;
+package com.midtrans.sdk.corekit.core.snap.model.pay.request;
 
 import com.google.gson.annotations.SerializedName;
